fix: Use INTEGER for featured column in PostgreSQL to fix type mismatch
- Change featured column from SMALLINT to INTEGER in PostgreSQL
- This fixes 'operator does not exist: boolean = integer' error
- Query now works with both PostgreSQL and MySQL
- Featured products will now display correctly

+++ setup.php
@@ -123,7 +123,7 @@ try {
             image_base64 TEXT DEFAULT NULL,
             image2_base64 TEXT DEFAULT NULL,
             image3_base64 TEXT DEFAULT NULL,
-            featured SMALLINT DEFAULT 0,
+            featured INTEGER DEFAULT 0,
             status VARCHAR(20) DEFAULT 'active',
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
